Add LocFilterName
Supporting loclist in QFilterName is easy; we just have to operate on
getloclist(0) instead of getqflist().

The filtering logic has been moved to s:FilterName, which now receives a
list as arg and returns the filtered list. This way, both QFilterName
and LocFilterName simply make use of s:FilterName, passing in the proper
list in each case.